Feat: Add price for Proposal struct

+++ src/response.rs
@@ -28,7 +28,7 @@ pub struct ProposalResponse {
   pub task_id: String,
   pub freelancer_id: String,
   pub milestones: Vec<MilestoneResponse>,
-  pub price: u16,
+  pub price: usize,
   pub accepted: bool,
 }
 
@@ -36,7 +36,7 @@ pub struct ProposalResponse {
 pub struct MilestoneResponse {
   pub description: String,
   pub deadline: String,
-  pub price: u16,
+  pub price: usize,
   pub status: String,
 }
